`@remotion/studio`: Share selected outline control state
