+++ packages/studio/src/components/SelectedOutlineEditingHandles.tsx
@@ -1,9 +1,4 @@
-import React, {useLayoutEffect, useMemo, useRef} from 'react';
-import {timelineSequenceNodePathToKey} from '../helpers/timeline-node-path-key';
-import {
-	useIsTimelineSequenceHovered,
-	useSetTimelineSequenceHover,
-} from '../state/timeline-sequence-hover';
+import React, {useMemo} from 'react';
 import {getSelectedOutlineControlLayout} from './selected-outline-control-layout';
 import type {SelectedOutline} from './selected-outline-geometry';
 import type {
@@ -20,6 +15,7 @@ import type {
 	TimelineSelection,
 	TimelineSelectionInteraction,
 } from './Timeline/TimelineSelection';
+import {useSelectedOutlineControlTarget} from './use-selected-outline-control-target';
 
 export const SelectedOutlineEditingHandles: React.FC<{
 	readonly dragging: boolean;
@@ -51,51 +47,14 @@ export const SelectedOutlineEditingHandles: React.FC<{
 	onSelect,
 	outline,
 }) => {
-	const setHoveredSequence = useSetTimelineSequenceHover();
-	const targetRef = useRef(layoutTarget);
-	useLayoutEffect(() => {
-		targetRef.current = layoutTarget;
-	}, [layoutTarget]);
-	const hoveredNodePathKey = useMemo(
-		() =>
-			layoutTarget === undefined
-				? null
-				: timelineSequenceNodePathToKey(
-						layoutTarget.nodePathInfo.sequenceSubscriptionKey,
-					),
-		[layoutTarget],
-	);
-	const hovered = useIsTimelineSequenceHovered(hoveredNodePathKey);
-	const controlTarget =
-		layoutTarget !== undefined && (layoutTarget.containsSelection || hovered)
-			? getLatestTargetByKey(layoutTarget.key)
-			: undefined;
+	const {controlTarget, hovered, onHoverChange} =
+		useSelectedOutlineControlTarget({
+			getLatestTargetByKey,
+			layoutTarget,
+		});
 	const controlLayout = useMemo(
 		() => getSelectedOutlineControlLayout(outline.points),
 		[outline.points],
-	);
-	const onHoverChange = React.useCallback(
-		(key: string | null) => {
-			setHoveredSequence((currentHover) => {
-				if (key !== null) {
-					const hoverTarget = targetRef.current;
-					if (hoverTarget === undefined || hoverTarget.key !== key) {
-						return currentHover;
-					}
-
-					return {
-						key,
-						nodePathKey: timelineSequenceNodePathToKey(
-							hoverTarget.nodePathInfo.sequenceSubscriptionKey,
-						),
-						source: 'canvas',
-					};
-				}
-
-				return currentHover?.source === 'canvas' ? null : currentHover;
-			});
-		},
-		[setHoveredSequence],
 	);
 	const onContextMenuOpen = React.useCallback(() => {
 		return getContextMenuOpenByKey(outline.key)?.() ?? false;
